+++ frontend/src/pages/detail/types/highlight.ts
@@ -1,0 +1,8 @@
+export interface HighlightType {
+  id: string;
+  color: string;
+  startXPath: string;
+  startOffset: number;
+  endXPath: string;
+  endOffset: number;
+}
