+++ frontend/src/routes/_bombom/articles.$articleId.tsx
@@ -11,6 +11,11 @@ import EmptyUnreadCard from '@/pages/detail/components/EmptyUnreadCard/EmptyUnre
 import MemoPanel from '@/pages/detail/components/MemoPanel/MemoPanel';
 import FloatingToolbar from '@/pages/detail/components/FloatingToolbar/FloatingToolbar';
 import NewsletterItemCard from '@/pages/detail/components/NewsletterItemCard/NewsletterItemCard';
+import { HighlightType } from '@/pages/detail/types/highlight';
+import {
+  restoreHighlight,
+  saveSelection,
+} from '@/pages/detail/utils/highlight';
 import { formatDate } from '@/utils/date';
 import ClockIcon from '#/assets/clock.svg';
 
@@ -18,213 +23,9 @@ export const Route = createFileRoute('/_bombom/articles/$articleId')({
   component: ArticleDetailPage,
 });
 
-/** 특정 노드의 XPath를 구하는 함수 */
-function getXPathForElement(node: Node, root: Node = document): string {
-  // TextNode일 경우 부모 요소로 이동
-  if (node.nodeType === Node.TEXT_NODE) {
-    node = node.parentNode!;
-  }
-  if (node === root) return '.';
-  const index =
-    Array.from(node.parentNode!.childNodes)
-      .filter((n) => n.nodeName === node.nodeName)
-      .indexOf(node as ChildNode) + 1;
-  return (
-    getXPathForElement(node.parentNode!, root) +
-    '/' +
-    node.nodeName.toLowerCase() +
-    `[${index}]`
-  );
-}
-
-/** XPath로 노드를 다시 찾는 함수 */
-function getElementByXPath(
-  xpath: string,
-  root: Document = document,
-): Node | null {
-  const result = document.evaluate(
-    xpath,
-    root,
-    null,
-    XPathResult.FIRST_ORDERED_NODE_TYPE,
-    null,
-  );
-  return result.singleNodeValue;
-}
-
-function getTextNodesInRange(range: Range): Text[] {
-  const textNodes: Text[] = [];
-
-  // === 선택된 범위가 단일 TextNode 내부 ===
-  if (
-    range.startContainer === range.endContainer &&
-    range.startContainer.nodeType === Node.TEXT_NODE
-  ) {
-    textNodes.push(range.startContainer as Text);
-    return textNodes;
-  }
-
-  // === 여러 노드일 경우 기존 TreeWalker 사용 ===
-  const walker = document.createTreeWalker(
-    range.commonAncestorContainer,
-    NodeFilter.SHOW_TEXT,
-    {
-      acceptNode(node) {
-        return range.intersectsNode(node)
-          ? NodeFilter.FILTER_ACCEPT
-          : NodeFilter.FILTER_REJECT;
-      },
-    },
-  );
-
-  let node: Node | null;
-  while ((node = walker.nextNode())) {
-    textNodes.push(node as Text);
-  }
-  return textNodes;
-}
-
-interface HighlightData {
-  id: string;
-  color: string;
-  startXPath: string;
-  startOffset: number;
-  endXPath: string;
-  endOffset: number;
-}
-
-function saveSelection(selection: Selection): HighlightData {
-  const range = selection.getRangeAt(0);
-  const container =
-    range.commonAncestorContainer.nodeType === Node.TEXT_NODE
-      ? range.commonAncestorContainer.parentElement!
-      : (range.commonAncestorContainer as Element);
-
-  const xpath = getXPathForElement(container);
-  const offsets = getOffsetsFromRangeIncludingMarks(container, range); // 변경
-  return {
-    startXPath: xpath,
-    startOffset: offsets.start,
-    endXPath: xpath,
-    endOffset: offsets.end,
-    color: '#FFEB3B',
-    id: crypto.randomUUID(),
-  };
-}
-
-function highlightNodeSegment(
-  node: Text,
-  start: number,
-  end: number,
-  color: string,
-  highlightId: string,
-) {
-  const parent = node.parentNode!;
-  const before = node.textContent!.slice(0, start);
-  const middle = node.textContent!.slice(start, end);
-  const after = node.textContent!.slice(end);
-
-  const mark = document.createElement('mark');
-  mark.style.backgroundColor = color;
-  mark.dataset.highlightId = highlightId;
-  mark.textContent = middle;
-
-  const frag = document.createDocumentFragment();
-  if (before) frag.appendChild(document.createTextNode(before));
-  frag.appendChild(mark);
-  if (after) frag.appendChild(document.createTextNode(after));
-
-  parent.replaceChild(frag, node);
-}
-
-function getOffsetsFromRangeIncludingMarks(container: Element, range: Range) {
-  let start = -1;
-  let end = -1;
-  let currentOffset = 0;
-
-  const walker = document.createTreeWalker(container, NodeFilter.SHOW_TEXT);
-
-  while (walker.nextNode()) {
-    const node = walker.currentNode as Text;
-    if (node === range.startContainer)
-      start = currentOffset + range.startOffset;
-    if (node === range.endContainer) end = currentOffset + range.endOffset;
-    currentOffset += node.textContent!.length;
-  }
-  return { start, end };
-}
-
-function createRangeFromOffsetsIncludingMarks(
-  container: Element,
-  start: number,
-  end: number,
-) {
-  const range = document.createRange();
-  const walker = document.createTreeWalker(container, NodeFilter.SHOW_TEXT);
-
-  let currentOffset = 0;
-  let startNode: Text | null = null;
-  let endNode: Text | null = null;
-  let startNodeOffset = 0;
-  let endNodeOffset = 0;
-
-  while (walker.nextNode()) {
-    const node = walker.currentNode as Text;
-    const len = node.textContent!.length;
-
-    // start node 찾기
-    if (!startNode && start >= currentOffset && start <= currentOffset + len) {
-      startNode = node;
-      startNodeOffset = start - currentOffset;
-    }
-
-    // end node 찾기
-    if (!endNode && end >= currentOffset && end <= currentOffset + len) {
-      endNode = node;
-      endNodeOffset = end - currentOffset;
-    }
-
-    currentOffset += len;
-  }
-
-  if (!startNode || !endNode) throw new Error('Offset 변환 실패');
-  range.setStart(startNode, startNodeOffset);
-  range.setEnd(endNode, endNodeOffset);
-
-  return range;
-}
-
-function restoreHighlight(data: HighlightData) {
-  const element = getElementByXPath(data.startXPath);
-  if (!element) return;
-
-  const range = createRangeFromOffsetsIncludingMarks(
-    element as Element,
-    data.startOffset,
-    data.endOffset,
-  );
-
-  // === 하이라이트 적용 ===
-  const highlightId = data.id || crypto.randomUUID();
-  const textNodes = getTextNodesInRange(range); // mark 제외 처리 가능
-  textNodes.forEach((node, index) => {
-    const isFirst = index === 0;
-    const isLast = index === textNodes.length - 1;
-
-    let start = 0;
-    let end = node.textContent!.length;
-    if (isFirst) start = range.startOffset;
-    if (isLast) end = range.endOffset;
-
-    if (start < end) {
-      highlightNodeSegment(node, start, end, data.color, highlightId);
-    }
-  });
-}
-
 function ArticleDetailPage() {
   const { articleId } = Route.useParams();
-  const [highlights, setHighlights] = useState<HighlightData[]>([]);
+  const [highlights, setHighlights] = useState<HighlightType[]>([]);
   const queryClient = useQueryClient();
   const [open, setOpen] = useState(true);
   const [memo, setMemo] = useState('');
@@ -338,8 +139,8 @@ function ArticleDetailPage() {
       </TodayArticlesWrapper>
       <FloatingToolbar
         onSave={(selection) => {
-          const highLightData = saveSelection(selection);
-          setHighlights((prev) => [...prev, highLightData]);
+          const HighlightType = saveSelection(selection);
+          setHighlights((prev) => [...prev, HighlightType]);
         }}
       />
       <MemoPanel
